swupdate: Split common parts into separate include
Split the swupdate_git.bb recipe into common include-file that can be used
to create swupdate recipes that are locked to a certain release without
duplicating recipe content.

The initscript bits are not part of the common include file since it is not
always desired to install an initscript (i.e when swupdate is not build to
go into background / no webserver).

Signed-off-by: Magnus Olsson <magnus@minimum.se>