swupdate class: add a way to uncondiionally files from deploy
Images in deploy directory are added to the cpio archive and
the resulting filename results from the concatenation of
the image name and MACHINE.

There are cases where files in deploy have no relationship with
the MACHINE - for example, if there are more as one DTB for the
same kernel. The DTB files result from the value set in
KERNEL_DEVICETREE and can have any value.

The new variable SWUPDATE_IMAGES_NOAPPEND_MACHINE is added
and evaluated by the class. If this is set for the specific image,
the file is simply added to the cpio archive without any changes
in the filename.

Signed-off-by: Stefano Babic <sbabic@denx.de>