swupdate_class: allow files with no fstype
Files in SWUPDATE_IMAGES with no fstype specified are currently skipped
when building the archive.  This changes the behaviour to include them
without appending an extension.

Together with the SWUPDATE_IMAGES_NOAPPEND_MACHINE flag this enables
inclusion of files from the deploy directory using the exact name in
SWUPDATE_IMAGES.

Signed-off-by: Mans Rullgard <mans@mansr.com>