use sstate staging to copy update image to deploy directory
Since morty YP uses the sstate also for the deploy directory. Using
this approach old images will be removed from deploy on a new build.
The same process is now used for updates images. As with normal images
the update images won't be cached in the sstate to reduce disk usage.

This change also resolves the issue, that update images are not
recreated after deleting the tmp directory and let bitbake regenerate
all unchanged recipes from sstate.

This is based on the original implementation in image.bbclass.

Signed-off-by: Christian Andersen <c.andersen@kostal.com>
Reviewed-by: Stefano Babic <sbabic@denx.de>