Fix build without network
bitbake runs git ls-remote if a recipe contains AUTOREV as SRCREV.
This forbids to build SWUpdate without network (setting BB_NO_NETWORK),
because the git version uses the last commit via AUTOREV.

Add a function to provide SRCREV for swupdate_git. This returns a fix
commit if the git version is not chosen, or AUTOREV if swupdate_git is selected.

Signed-off-by: Stefano Babic <sbabic@denx.de>
Reported-by: Diaz de Grenu, Jose <Jose.DiazdeGrenu@digi.com>