swupdate,recuegui: get rid of AUTOREV
It's best practice in OE to have fixed revisions in _git recipes.
The main reason is that AUTOREV requires a connection to the server during
parsing. This will fail if no internet access is possible.

If AUTOREV is desired it should be put in local.conf via:

SRCREV_pn-PN = "${AUTOREV}"

This patch replaces all current uses of AUTOREV with the latest available
git revision.

Signed-off-by: Pascal Bach <pascal.bach@siemens.com>
Acked-by: Stefano Babic <sbabic@denx.de>