Extend Allow expansion of variables in sw-description
Variables are just matched one for each line in sw-description. The
patch extend the behavior and searches recursively in each line for all
matches of the substituion pattern @@variable@@.

This adds supports for variable's flags, too. The following match is
also resolved:

	@@variable[name-of-flag]@@

This is useful to evaluate SWUpdate's flags as the image extension.

Signed-off-by: Stefano Babic <sbabic@denx.de>