Check for fetched URLs before adding to SWU
There are cases where the fetched URLs are retrieved from a repo. Files
cannot be copied as they are and must be evaluated in the recipe. IOne
use case is if the sw-description is fetched from a git repo instead of
beeing stored on a local file. The check vrifies that the fetched URL is
a file before copying it.

Signed-off-by: Stefano Babic <sbabic@denx.de>