mtd-utils: Use include dir that upstream suggests
This prepares to move mtd headers in /usr/include instead of the mtd/
namespace used by kernel. This will be synchronized with future dev
packages for mtd-utils that will be merged into Linux distros (Debian).
For a while, the OE recipe will still contain links to the old path to
be compatible with previous SWUpdate release. The links should be
dropped in future.

Signed-off-by: Stefano Babic <sbabic@denx.de>
CC: Bastian Germann <bage@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>