swupdate-lualoader: Add recipe
Adds a recipe for a simple swudpate_handlers.lua script that loads lua
files from a directory. This allows other recipes to drop lua handler
scripts in the specified location and swupdate will find them instead of
having to explicitly list all the files to be loaded in a custom
swupdate_handlers.lua script.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>
Reviewed-by: stefano Babic <sbabic@denx.de>