Fix breakage if encryption is not set
Commit introduces the encryption into the swupdate class, but it breaks
builds if encryption is not used at all. Fix this and rework function
to retrieve key from a file.

Replace os.system() with subprocess.run() - this should be done for other
occurrencies od os.system() that is declared obsolete.

Signed-off-by: Stefano Babic <sbabic@denx.de>