swupdate.bbclass: fix SWUPDATE_IMAGES_NOAPPEND_MACHINE parsing
The SWUPDATE_IMAGES_NOAPPEND_MACHINE check is incorrect and always takes
the 'elif' branch. This is because the allowed values for
SWUPDATE_IMAGES_NOAPPEND_MACHINE are "0", "1" or unset. The first two
values are strings, and getVarFlag() returns them as strings, but the code
checks for boolean.

Keep the allowed values to "0", "1" or unset as per the documentation and
to keep backward compatibility, and fix by comparing with strings instead
of booleans.

Fixes: b6534bbc2781 ("swupdate: simplify find images added to swu")
Signed-off-by: Luca Ceresoli <luca@lucaceresoli.net>
Acked-by: Stefano Babic <sbabic@denx.de>