Create a class to generate SWU from image
It is currently required a separate recipe to build the SWU from an
image. This is very flexible but it appears overkilling in use cases
where the only artifact is the rootfs. Create a separate class that can
be inherited replacing swupdate.bbclass and that build a SWU from an
image recipe.

Signed-off-by: Stefano Babic <sbabic@denx.de>