swupdate-common: add find bitbake variables
The new function is supposed to read all variables from sw-description
file and add them to the vardeps of the do_swuimage task. Bitbake
cannot know that the do_swuimage task which evaluates the templated
sw-description file needs to be executed if a variable which is
refered by the sw-description file but not by the recipe itself.

Using this function should be optional. That's why it is not called yet.

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>