swupdate-image: use image link name
Use IMAGE_LINK_NAME instead of IMAGE_BASENAME. IMAGE_BASENAME contains a
timestamp which causes some sporadic build failures. The problem is that
if the image files have been built by a previous bitbake call and bitbake
skips the task to build them again swupdate fails with searching for images
with the timestamp of the current build job in the file name. The link
refers to the latest available build in any case.

Setting S to WORKDIR should not be needed. It's the default from
bitbake.conf.

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>