swupdate-common: Fix signing subprocess call
The subprocess call in prepare_sw_description to sign firmware was
modifed from os.system to subprocess.run commit a5526cc.  If multiple
commands are chained in SWUPDATE_SIGN_TOOL with ";", "&&", or "||", only
the first command is now evaluated.
This behavior was readded by setting shell=True in the subprocess call
along with joining args to a single string.

Signed-off-by: Colin McAllister <colin.mcallister@garmin.com>