class: add function that signs and returns hash of artifact
This adds swupdate_sign_ec521_sha512 - it can be used to report and store the
hash after signing an artifact. Used to verify the integrity of a
component by the device during boot. This function signs a file with
secp521r1.

There are a lot of ways to sign, and they can be easily added later. The
rule is to add a function with the name swupdate_sign_<signing
algotithm>, and if the function needs OE variables, they should be
created with the name rule SWUPDATE_<algotithm>_<varname>.

Signed-off-by: Stefano Babic <sbabic@denx.de>