class: rework autoversion to be more generic
Drop special parsing with @SWU_VERSION, not required anymore, because it
is possible to call a function by setting the name inside
sw-description. Simplify the workflow and the search pattern. Drop
limitation that a package should be part of SWU, and look up for any
variable belonging to any package. Drop the association between filename
and package name, because it is in most cases wrong and it is replaced
by the explicit name of the package in sw-description.

Old syntax:

	version = "@SWU_AUTO_VERSION"; // package name = filename
		or
	version = "@SWU_AUTO_VERSION:<package name>"
		or
	version = "@SWU_AUTO_VERSION:<package name>@<variable>"

New syntax:

	<any attribute> = "$swupdate_auto_versions(<package name>)";
		or
	<any attribute> = "$swupdate_auto_versions(<package name>@<variable>)";

Signed-off-by: Stefano Babic <sbabic@denx.de>
CC: Thomas Haemmerle <thomas.haemmerle@leica-geosystems.com>