swupdate-usb: add direct dependency on swupdate.service
There is a true dependecy of swupdate-usb service on swupdate
service which is not present in a proper way. If swupdate is
not running for some reason, the swupdate-usb service will fail
as swupdate-client requires the host socket to be present which
is provided by host.
In current implementation swupdate-usb has a dependedency on
swupdate-progress which does not have a "Requires" dependency
on swupdate. This results in a failure if swupdate service is not
running and swupdate-usb tries to invoke swupdate-client.
The swupdate-sb must have a direct dependency on swupdate service
because we might have systems where one does not want to monitor
progress. Therefore any changes in current dependency chain will
result in failure for swupdate.
Therefore a "Require" dependency on swupdate service must be
present in swupdate-usb. swsupdate-usb does not require
swupdate-progress service to be present or running for a successful
update but it does need swupdate.service to be up and running to
run. This is makes it a true dependency.
This commit changes swupdate-usb to require swsupdate service to
have a proper and right dedpendency.

Signed-off-by: Muhammad Hamza <muhammad_hamza@mentor.com>
Acked-by: Stefano Babic <sbabic@denx.de>