swupdate: package libswupdate into separate package
The libswupdate shared library produced by the contents of the ipc
subdirectory can be used by other packages to communicate with SWUpdate.
The library is embedded into the base package which creates two
problems:

- A recipe / program cannot depend on libswupdate without depending on
  swupdate
- The complete set of swupdate licenses seems to apply to the program
  including strong copyleft GPL-2.0-only instead only LPGL-2.1-or-later

As documented in the source code swupdate-ipc is purely
LGPL-2.1-or-later.

Existing recipes / programs which need libswupdate but only depend on
swupdate will still get libswupdate through a runtime recommendation.

Signed-off-by: Ricardo Simoes <ricardo.simoes@pt.bosch.com>
Signed-off-by: Mark Jonas <mark.jonas@de.bosch.com>