swupdate-common: don't change files in ${S}
Since do_unpack() is only executed whenever files in ${SRC_URI} have
changed, the do_swuimage() can end up working with a swu-description
file that is the output of a previous execution of do_swuimage(). This
is because do_unpack() copies swu-description to ${S} and do_swuimage()
uses ${S} for both input and output.

The effect of using ${S} for both input and output is that placeholdes
in swu-description is only replaced on the first invocation of
do_swuimage(). This breaks in case of checksums, as the input artifact
might change without do_unpack() being called again.

Fix this by using a separate ${B}, and thereby avoid changing the
swu-description in ${S}.