recipes-support: swupdate.sh: improve handling of quoted arguments
Attempting to set the downloader arguments via SWUPDATE_EXTRA_ARGS in
the swupdate systemd service file fails:

    [Service]
    Environment="SWUPDATE_EXTRA_ARGS=-d '-r 3'"

Due to the way the swupdate binary is invoked, the resulting command
line arguments for swupdate end up as:

  argv[1] = "-v"     # set in the swupdate.sh script
  argv[2] = "-d"     # here comes SWUPDATE_EXTRA_ARGS
  argv[3] = "'-r"    # not the expected behaviour from quotes
  argv[4] = "3'"

If swupdate is invoked with "eval exec" instead of "exec", we get the
expected behaviour from quotes:

  argv[1] = "-v"     # set in the swupdate.sh script
  argv[2] = "-d"     # here comes SWUPDATE_EXTRA_ARGS
  argv[3] = "-r 3"   # expected behaviour from quotes

The explicit invocation of swupdate when SWUPDATE_WEBSERVER_ARGS or
SWUPDATE_SURICATTA_ARGS are set, is probably for the same reason.
Simplify these as "eval exec" will also give the expected behaviour for
these arguments.